# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "easyOrder 3.0",
    "description": "Sistema de Gestão de Pedidos - Tech Challenge (3) da Pos Tech SOAT ",
    "version": "1.0.0"
  },
  "host": "",
  "paths": {
...
    "/preparacao/pedido/proximo": {
      "get": {
        "tags": [
          "Preparação"
        ],
        "summary": "Buscar o próximo Pedido",
        "description": "Busca o próximo pedido para preparação<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Pedido encontrado"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Pedido encontrado"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhum pedido encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Nenhum pedido encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Nenhum pedido encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/preparacao/pedido/{pedidoId}/iniciar-preparacao": {
      "put": {
        "tags": [
          "Preparação"
        ],
        "summary": "Iniciar Preparação de Pedido",
        "description": "Inicia a preparação de um pedido<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do pedido",
            "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
          }
        ],
        "responses": {
          "200": {
            "description": "Preparação do pedido iniciada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Preparação do pedido iniciada com sucesso"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Preparação do pedido iniciada com sucesso"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/preparacao/pedido/{pedidoId}/finalizar-preparacao": {
      "put": {
        "tags": [
          "Preparação"
        ],
        "summary": "Finalizar Preparação de Pedido",
        "description": "Finaliza a preparação de um pedido<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do pedido",
            "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
          }
        ],
        "responses": {
          "200": {
            "description": "Preparação do pedido finalizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Preparação do pedido finalizada com sucesso"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Preparação do pedido finalizada com sucesso"
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/preparacao/pedido/{pedidoId}/entregar": {
      "put": {
        "tags": [
          "Preparação"
        ],
        "summary": "Entregar Pedido",
        "description": "Entrega um pedido preparado<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do pedido",
            "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido entregue com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Pedido entregue com sucesso."
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Pedido entregue com sucesso."
                    },
                    "pedido": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Pedido não encontrado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Pedido": {
        "type": "object",
        "properties": {
          "id": {
            "schema": {
              "$ref": "#/definitions/ID"
            }
          },
          "data": {
            "type": "string",
            "example": "2021-10-10T15:00:00.000Z"
          },
          "cliente_identificado": {
            "type": "boolean",
            "example": true
          },
          "clienteId": {
            "$ref": "#/definitions/ID"
          },
          "status": {
            "type": "string",
            "example": "EM_ABERTO"
          },
          "pagamentoStatus": {
            "type": "string",
            "example": "PENDENTE"
          },
          "combos": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Combo"
            }
          },
          "valorTotal": {
            "type": "number",
            "example": 80
          }
        }
      }
    }
  }
}